Improve castSpellOnNpc API: accept spell names instead of raw IDs
Addresses #24 - the castSpellOnNpc API was confusing because it required
opaque numeric spell component IDs with no guidance on valid values.

Changes:
- Add sdk/spells.ts as the canonical source for spell constants, exported
  as `Spells` with a `SpellName` type derived from its keys
- Update castSpellOnNpc, sendSpellOnNpc, and sendSpellOnItem to accept
  `SpellName | number` (e.g. 'WIND_STRIKE' or 1152), following the same
  pattern as sendTogglePrayer which accepts PrayerName | number
- Re-export Spells and SpellName from actions.ts for convenient imports
- Replace inline Spells definition in save-generator.ts with re-export
- Remove duplicate Spells constant from scripts/magic/script.ts
- Document all spell IDs in API.md with a new Constants section

Closes #24
