+++ sdk/test/utils/save-generator.ts
@@ -108,42 +108,8 @@ export const Items = {
     BIG_BONES: 532,
 };
 
-// Spell component IDs (from content/pack/interface.pack)
-export const Spells = {
-    // Combat spells
-    WIND_STRIKE: 1152,
-    CONFUSE: 1153,
-    WATER_STRIKE: 1154,
-    ENCHANT_LVL1: 1155,  // Sapphire
-    EARTH_STRIKE: 1156,
-    WEAKEN: 1157,
-    FIRE_STRIKE: 1158,
-    WIND_BOLT: 1160,
-    CURSE: 1161,
-    LOW_ALCHEMY: 1162,
-    WATER_BOLT: 1163,
-    VARROCK_TELEPORT: 1164,
-    ENCHANT_LVL2: 1165,  // Emerald
-    EARTH_BOLT: 1166,
-    LUMBRIDGE_TELEPORT: 1167,
-    FIRE_BOLT: 1169,
-    FALADOR_TELEPORT: 1170,
-    WIND_BLAST: 1172,
-    SUPERHEAT: 1173,
-    CAMELOT_TELEPORT: 1174,
-    WATER_BLAST: 1175,
-    ENCHANT_LVL3: 1176,  // Ruby
-    EARTH_BLAST: 1177,
-    HIGH_ALCHEMY: 1178,
-    ENCHANT_LVL4: 1180,  // Diamond
-    FIRE_BLAST: 1181,
-    WIND_WAVE: 1183,
-    WATER_WAVE: 1185,
-    ENCHANT_LVL5: 1187,  // Dragonstone
-    EARTH_WAVE: 1188,
-    FIRE_WAVE: 1189,
-    BIND: 1572,
-};
+// Re-export spell constants from the canonical source
+export { Spells } from '../../spells';
 
 // Skill indices
 export const Skills = {
